state: move restore functionality into its own file.
